Crate rmathlib

A Rust port of R’s C Library of Special Functions.

Functions

bd0

Calculates a stable deviance part using a method that reduces relative error

bratio

Evaluation of the Incomplete Beta function I_x(a,b)

chebyshev_eval

chebyshev_eval evaluates the n-term Chebyshev series a at x.

chebyshev_init

chebyshev_init determines the number of terms for the double precision orthogonal series dos needed to insure the error is no larger than eta. Ordinarily eta will be chosen to be one-tenth machine precision.

cospi

Calculates the cosine of a number given in multiples of π (pi).

dgamma

AUTHOR Catherine Loader, catherine@research.bell-labs.com. October 23, 2000.

dnorm

dpois

checks argument validity and calls dpois_raw().

ebd0

Compute x * log (x / M) + (M - x) aka -x * log1pmx ((M - x) / x)

gammafn

This function computes the value of the gamma function.

i1mach

lbeta

This function returns the value of the log beta function

lgammacor

Compute the log gamma correction factor for x >= 10 so that

lgammafn

The function lgammafn computes log|gamma(x)|. The function lgammafn_sign in addition assigns the sign of the gamma function to the address in the second argument if this is not NULL.

lgammafn_sign

Compute the log gamma function and its sign.

log1pmx

Accurate calculation of log(1+x)-x, particularly for small x.

logspace_add

Compute the log of a sum from logs of terms, i.e., log(exp(logx) + exp(logy)) without causing overflows and without throwing away large handfuls of accuracy.

pbeta

Returns distribution function of the beta distribution. ( = The incomplete beta ratio I_x(p,q) ).

pgamma

Computes the distribution function for the gamma distribution with shape parameter alph and scale parameter scale.

pnorm

pt

The PDF of the Student’s t-distribution.

qnorm

sinpi

Calculates the sinus of a number given in multiples of π (pi).

stirlerr

Computes the log of the error term in Stirling’s formula.

tanpi

Calculates the tangent of a number given in multiples of π (pi).