Function rmathlib::dgamma

pub fn dgamma(x: f64, shape: f64, scale: f64, give_log: bool) -> f64

AUTHOR Catherine Loader, catherine@research.bell-labs.com. October 23, 2000.

Merge in to R: Copyright (C) 2000-2019 The R Core Team Copyright (C) 2004-2019 The R Foundation

This program is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation; either version 2 of the License, or (at your option) any later version.

This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.

You should have received a copy of the GNU General Public License along with this program; if not, a copy is available at https://www.R-project.org/Licenses/ Computes the density of the Gamma distribution.

_ 1/s (x/s)^{a-1} exp(-x/s) p(x;a,s) = ———————– _ (a-1)!

where ‘s’ is the scale (= 1/lambda in other parametrizations) and ‘a’ is the shape parameter ( = alpha in other contexts).